refactor(parse-searchset): remove dead code flagged by DeepScan
DeepScan reported CONSTANT_CONDITION warnings for `if (parsed)` guards
that are always true because the extraction functions never return null.

Changes to src/utils/parse-searchset-bundles.ts:
- Remove `| null` from `extractObservation` return type
- Inline `extractQuestionnaireResponse` call in `parseQRBundle`, remove guard
- Inline `extractCondition` call in `parseConditionBundle`, remove guard
- Inline `extractObservation` call in `parseObservationBundle`, remove guard

Test split: separated the oversized test file into per-concern files
to stay under the 300-line ESLint limit:
- src/utils/__tests__/parse-searchset-bundles.test.ts (QR tests only)
- src/utils/__tests__/parse-searchset-bundles.condition.test.ts (Condition tests)
- src/utils/__tests__/parse-searchset-bundles.observation.test.ts (Observation tests)
- src/utils/__tests__/parse-searchset-bundles.edge.test.ts (edge-case tests)
